LayoutTests/http/tests/site-isolation/inspector/console/message-from-iframe-expected.txt

CONSOLE MESSAGE: Hello from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
CONSOLE MESSAGE: Warning from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
CONSOLE MESSAGE: Error from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
CONSOLE MESSAGE: iframe 0 is loaded in http://127.0.0.1:8000/site-isolation/inspector/console/message-from-iframe.html
CONSOLE MESSAGE: Hello from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
CONSOLE MESSAGE: Warning from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
CONSOLE MESSAGE: Error from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
CONSOLE MESSAGE: iframe is loaded in iframe http://127.0.0.1:8000/site-isolation/inspector/console/resources/embedded-same-origin.html
CONSOLE MESSAGE: iframe 2 is loaded in http://127.0.0.1:8000/site-isolation/inspector/console/message-from-iframe.html

== Running test suite: MessageFromIFrame
-- Running test case: MessageFromIFrame.SameOriginIFrame
Got message from frame target: Hello from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
Got message from frame target: Warning from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
Got message from frame target: Error from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
Got message from frame target: iframe 0 is loaded in http://127.0.0.1:8000/site-isolation/inspector/console/message-from-iframe.html

-- Running test case: MessageFromIFrame.GrandChildSameOriginIFrameInSameOrigin
Got message from frame target: Hello from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
Got message from frame target: Warning from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
Got message from frame target: Error from http://127.0.0.1:8000/site-isolation/inspector/console/resources/console-messages.html
Got message from frame target: iframe is loaded in iframe http://127.0.0.1:8000/site-isolation/inspector/console/resources/embedded-same-origin.html
Got message from frame target: iframe 2 is loaded in http://127.0.0.1:8000/site-isolation/inspector/console/message-from-iframe.html
